% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_series.R
\name{dexp_series}
\alias{dexp_series}
\title{pdf for exponential series.}
\usage{
dexp_series(t, rates, log = FALSE)
}
\arguments{
\item{t}{series system lifetime}

\item{rates}{rate parameters for exponential component lifetimes}

\item{log}{return the log of the pdf}
}
\value{
Density values for the exponential series distribution.
}
\description{
pdf for exponential series.
}
\examples{
rates <- c(0.5, 0.3, 0.2)
dexp_series(1.0, rates)
dexp_series(c(0.5, 1.0, 2.0), rates, log = TRUE)
}
