% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_series_masked_component_cause.R
\name{md_cand_sampler}
\alias{md_cand_sampler}
\title{Sample candidate sets for systems with unobserved components.}
\usage{
md_cand_sampler(df, prob = "q", candset = "x")
}
\arguments{
\item{df}{(masked) data frame}

\item{prob}{column prefix for component probabilities, defaults to
\code{q}, e.g., \verb{q1, q2, q3}.}

\item{candset}{column prefix for candidate sets (as Boolean matrix),
defaults to \code{x}, e.g., \verb{x1, x2, x3}.}
}
\value{
Data frame with added Boolean candidate set columns
(e.g., \code{x1, x2, ..., xm}).
}
\description{
Candidate set generator. Requires columns for component probabilities
e.g., \verb{q1,...,qm} where \code{qj} is the probability that the jth component
will be in the corresponding candidate set generated for that observation
in the \code{md} table.
}
\examples{
# Generate component lifetimes
set.seed(42)
mat <- matrix(rexp(9, rate = rep(c(0.5, 0.3, 0.2), each = 3)),
              nrow = 3, ncol = 3)
df <- md_encode_matrix(mat, "t")
df$t <- apply(mat, 1, min)
df$delta <- TRUE
df <- md_bernoulli_cand_c1_c2_c3(df, p = 0.5)
md_cand_sampler(df)
}
